Force XCB backend to prevent crashes on Wayland
